Split up background.js
