+++ extension/background-script/config.js
@@ -1,0 +1,42 @@
+// @ts-check
+/// <reference path="../../types/chrome.d.ts" />
+/// <reference path="../../types/index.js" />
+
+/** @type {Intl.DateTimeFormatOptions} */
+export const timeFormat = {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+    hour12: true
+}
+
+/** @type {Object<string, { id: string, js: string[], matches: string[], excludeMatches: string[], permissions: chrome.runtime.ManifestPermissions[] }>} */
+export const PLATFORM_CONFIGS = {
+    "google_meet": {
+        id: "content-google-meet",
+        js: ["content-google-meet.js"],
+        matches: ["https://meet.google.com/*"],
+        excludeMatches: ["https://meet.google.com/", "https://meet.google.com/landing"],
+        permissions: ["notifications"]
+    },
+    "teams": {
+        id: "content-teams",
+        js: ["content-teams.js"],
+        matches: ["https://teams.live.com/*", "https://teams.microsoft.com/*", "https://teams.cloud.microsoft/*"],
+        excludeMatches: [],
+        permissions: ["notifications"]
+    },
+    "zoom": {
+        id: "content-zoom",
+        js: ["content-zoom.js"],
+        matches: ["https://*.zoom.us/*"],
+        excludeMatches: [],
+        permissions: ["notifications", "declarativeNetRequestWithHostAccess"]
+    }
+}
+
+export const ALARM_NAME = "dailyPermissionCheck"
+export const INTERVAL_IN_MINUTES = 24 * 60 // 24 hours
+
