--- extension/background.js
@@ -1,1187 +1,0 @@
-// @ts-check
-/// <reference path="../types/chrome.d.ts" />
-/// <reference path="../types/index.js" />
-
-/** @type {Intl.DateTimeFormatOptions} */
-const timeFormat = {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: true
-}
-
-/** @type {Object<string, { id: string, js: string[], matches: string[], excludeMatches: string[], permissions: chrome.runtime.ManifestPermissions[] }>} */
-const PLATFORM_CONFIGS = {
-    "google_meet": {
-        id: "content-google-meet",
-        js: ["content-google-meet.js"],
-        matches: ["https://meet.google.com/*"],
-        excludeMatches: ["https://meet.google.com/", "https://meet.google.com/landing"],
-        permissions: ["notifications"]
-    },
-    "teams": {
-        id: "content-teams",
-        js: ["content-teams.js"],
-        matches: ["https://teams.live.com/*", "https://teams.microsoft.com/*", "https://teams.cloud.microsoft/*"],
-        excludeMatches: [],
-        permissions: ["notifications"]
-    },
-    "zoom": {
-        id: "content-zoom",
-        js: ["content-zoom.js"],
-        matches: ["https://*.zoom.us/*"],
-        excludeMatches: [],
-        permissions: ["notifications", "declarativeNetRequestWithHostAccess"]
-    }
-}
-
-const ALARM_NAME = "dailyPermissionCheck"
-const INTERVAL_IN_MINUTES = 24 * 60 // 24 hours
-
-
-chrome.runtime.onMessage.addListener(function (messageUnTyped, sender, sendResponse) {
-    const message = /** @type {ExtensionMessage} */ (messageUnTyped)
-    console.log(message.type)
-
-    if (message.type === "new_meeting_started") {
-        // Saving current tab id, to download transcript when this tab is closed
-        chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
-            const tabId = tabs[0].id
-            chrome.storage.local.set({ meetingTabId: tabId }, function () {
-                console.log("Meeting tab id saved")
-            })
-        })
-    }
-
-    if (message.type === "meeting_ended") {
-        // Prevents double downloading of transcript from tab closed event listener. Also prevents available update from being applied, during meeting post processing.
-        chrome.storage.local.set({ meetingTabId: "processing" }, function () {
-            console.log("Meeting tab id set to processing meeting")
-
-            processLastMeeting()
-                .then(() => {
-                    /** @type {ExtensionResponse} */
-                    const response = { success: true }
-                    sendResponse(response)
-                })
-                .catch((error) => {
-                    // Fails with error codes: 009, 010, 011, 012, 013, 014
-                    const parsedError = /** @type {ErrorObject} */ (error)
-
-                    /** @type {ExtensionResponse} */
-                    const response = { success: false, message: parsedError }
-                    sendResponse(response)
-                })
-                .finally(() => {
-                    // setTimeout(() => {
-                    //     checkPermissionsAndOpenMeetingsPage()
-                    // }, 10000)
-                    clearTabIdAndApplyUpdate()
-                })
-        })
-    }
-
-    if (message.type === "download_transcript_at_index") {
-        if ((typeof message.index === "number") && (message.index >= 0)) {
-            // Download the requested item
-            downloadTranscript(message.index, false)
-                .then(() => {
-                    /** @type {ExtensionResponse} */
-                    const response = { success: true }
-                    sendResponse(response)
-                })
-                .catch((error) => {
-                    // Fails with error codes: 009, 010
-                    const parsedError = /** @type {ErrorObject} */ (error)
-
-                    /** @type {ExtensionResponse} */
-                    const response = { success: false, message: parsedError }
-                    sendResponse(response)
-                })
-        }
-        else {
-            /** @type {ExtensionResponse} */
-            const response = { success: false, message: { errorCode: "015", errorMessage: "Invalid index" } }
-            sendResponse(response)
-        }
-    }
-
-    if (message.type === "post_webhook_at_index") {
-        if ((typeof message.index === "number") && (message.index >= 0)) {
-            // Handle webhook retry
-            postTranscriptToWebhook(message.index)
-                .then(() => {
-                    /** @type {ExtensionResponse} */
-                    const response = { success: true }
-                    sendResponse(response)
-                })
-                .catch(error => {
-                    // Fails with error codes: 009, 010, 011, 012
-                    const parsedError = /** @type {ErrorObject} */ (error)
-
-                    console.error("Webhook retry failed:", parsedError)
-                    /** @type {ExtensionResponse} */
-                    const response = { success: false, message: parsedError }
-                    sendResponse(response)
-                })
-        }
-        else {
-            /** @type {ExtensionResponse} */
-            const response = { success: false, message: { errorCode: "015", errorMessage: "Invalid index" } }
-            sendResponse(response)
-        }
-    }
-
-    if (message.type === "recover_last_meeting") {
-        recoverLastMeeting().then((message) => {
-            /** @type {ExtensionResponse} */
-            const response = { success: true, message: message }
-            sendResponse(response)
-        })
-            .catch((error) => {
-                // Fails with error codes: 009, 010, 011, 012, 013, 014
-                const parsedError = /** @type {ErrorObject} */ (error)
-
-                /** @type {ExtensionResponse} */
-                const response = { success: false, message: parsedError }
-                sendResponse(response)
-            })
-    }
-
-    if (message.type === "get_platform_enablement_status") {
-        /** @type {Platform | Platform[]} */
-        let platform = message.platform || "google_meet"
-
-        getContentScriptStatus(platform)
-            .then((status) => {
-                /** @type {ExtensionResponse} */
-                const response = {
-                    success: true,
-                    message: status
-                }
-                sendResponse(response)
-            })
-            .catch((error) => {
-                const parsedError = /** @type {ErrorObject} */ (error)
-                /** @type {ExtensionResponse} */
-                const response = { success: false, message: parsedError }
-                sendResponse(response)
-            })
-    }
-
-    if (message.type === "get_platform_permission_status") {
-        /** @type {Platform | Platform[]} */
-        let platform = message.platform || "google_meet"
-
-        getPermissionStatus(platform)
-            .then((status) => {
-                /** @type {ExtensionResponse} */
-                const response = {
-                    success: true,
-                    message: status
-                }
-                sendResponse(response)
-            })
-            .catch((error) => {
-                const parsedError = /** @type {ErrorObject} */ (error)
-                /** @type {ExtensionResponse} */
-                const response = { success: false, message: parsedError }
-                sendResponse(response)
-            })
-    }
-
-    if ((message.type === "enable_platform")) {
-        /** @type {Platform | Platform[]} */
-        let platform = message.platform || "google_meet"
-
-        requestPlatformPermission(platform).then(() => {
-            // After permissions are granted, register both the scripts and the redirect rule for zoom
-            const promises = [registerContentScript(platform)]
-
-            const platformsArray = Array.isArray(platform) ? platform : [platform]
-            const hasZoom = platformsArray.includes("zoom")
-
-            if (hasZoom) {
-                promises.push(registerZoomRedirect())
-            }
-
-            Promise.all(promises).then((results) => {
-                /** @type {ExtensionResponse} */
-                const response = { success: true, message: results[0] }
-                sendResponse(response)
-            }).catch((error) => {
-                // Fails with error codes: not defined
-                const parsedError = /** @type {ErrorObject} */ (error)
-
-                /** @type {ExtensionResponse} */
-                const response = { success: false, message: parsedError }
-                sendResponse(response)
-            })
-        })
-            .catch((error) => {
-                // Fails with error codes: not defined
-                const parsedError = /** @type {ErrorObject} */ (error)
-
-                /** @type {ExtensionResponse} */
-                const response = { success: false, message: parsedError }
-                sendResponse(response)
-            })
-    }
-
-    if (message.type === "disable_platform") {
-        /** @type {Platform | Platform[]} */
-        let platform = message.platform || "google_meet"
-
-        // To disable, we simply unregister the content scripts and any redirect rules
-        const promises = [deregisterContentScript(platform)]
-
-        const platformsArray = Array.isArray(platform) ? platform : [platform]
-        const hasZoom = platformsArray.includes("zoom")
-
-        if (hasZoom) {
-            promises.push(deregisterZoomRedirect())
-        }
-
-        Promise.all(promises).then((results) => {
-            /** @type {ExtensionResponse} */
-            const response = { success: true, message: results[0] }
-            sendResponse(response)
-        }).catch((error) => {
-            const parsedError = /** @type {ErrorObject} */ (error)
-
-            /** @type {ExtensionResponse} */
-            const response = { success: false, message: parsedError }
-            sendResponse(response)
-        })
-    }
-
-    if (message.type === "open_popup") {
-        /** @type {Platform} */
-
-        openExtensionPopup().then((message) => {
-            /** @type {ExtensionResponse} */
-            const response = { success: true, message: message }
-            sendResponse(response)
-        }).catch((error) => {
-            const parsedError = /** @type {ErrorObject} */ (error)
-
-            /** @type {ExtensionResponse} */
-            const response = { success: false, message: parsedError }
-            sendResponse(response)
-        })
-    }
-
-
-    return true
-})
-
-// Download last meeting if meeting tab is closed
-chrome.tabs.onRemoved.addListener(function (tabId) {
-    chrome.storage.local.get(["meetingTabId"], function (resultLocalUntyped) {
-        const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-
-        if (tabId === resultLocal.meetingTabId) {
-            console.log("Successfully intercepted tab close")
-
-            // Prevent misfires of onRemoved until next meeting. Also prevents available update from being applied, during meeting post processing.
-            chrome.storage.local.set({ meetingTabId: "processing" }, function () {
-                console.log("Meeting tab id set to processing meeting")
-
-                processLastMeeting().finally(() => {
-                    // setTimeout(() => {
-                    //     checkPermissionsAndOpenMeetingsPage()
-                    // }, 10000)
-                    clearTabIdAndApplyUpdate()
-                })
-            })
-        }
-    })
-})
-
-// Listen for extension updates
-chrome.runtime.onUpdateAvailable.addListener(() => {
-    // Check if there is an active meeting
-    chrome.storage.local.get(["meetingTabId"], function (resultUntyped) {
-        const result = /** @type {ResultLocal} */ (resultUntyped)
-
-        if (result.meetingTabId) {
-            // There is an active meeting(values: tabId or processing), defer the update
-            chrome.storage.local.set({ isDeferredUpdatedAvailable: true }, function () {
-                console.log("Deferred update flag set")
-            })
-        } else {
-            // No active meeting, apply the update immediately. Meeting tab id is nullified only post meeting operations are done, so no race conditions.
-            console.log("No active meeting, applying update immediately")
-            chrome.runtime.reload()
-        }
-    })
-})
-
-// Register content scripts and Zoom redirect whenever runtime permission change—mostly serves as a backup for changes made outside the UI.
-chrome.permissions.onAdded.addListener((event) => {
-    // Prevent competing with explicit content script registrations
-    setTimeout(() => {
-        /** @type {Platform[]} */
-        const platforms = ["google_meet", "teams", "zoom"]
-        /** @type {Promise<any>[]} */
-        const deRegisterPromises = [deregisterContentScript(platforms), deregisterZoomRedirect()]
-
-        Promise.all(deRegisterPromises)
-            .then(() => {
-                console.log("De-registered all content scripts. Starting re-registration...")
-
-            })
-
-        reRegisterContentScripts()
-    }, 2000)
-})
-
-
-chrome.runtime.onInstalled.addListener(() => {
-    // Set defaults values
-    chrome.storage.sync.get(["autoPostWebhookAfterMeeting", "autoDownloadFileAfterMeeting", "operationMode", "webhookBodyType", "webhookUrl", "wantGoogleMeet", "wantTeams", "wantZoom"], function (resultSyncUntyped) {
-        const resultSync = /** @type {ResultSync} */ (resultSyncUntyped)
-
-        console.log(resultSync.wantTeams)
-
-        chrome.storage.sync.set({
-            autoPostWebhookAfterMeeting: resultSync.autoPostWebhookAfterMeeting === false ? false : true,
-            autoDownloadFileAfterMeeting: resultSync.autoDownloadFileAfterMeeting === false ? false : true,
-            operationMode: resultSync.operationMode === "manual" ? "manual" : "auto",
-            webhookBodyType: resultSync.webhookBodyType === "advanced" ? "advanced" : "simple",
-            wantGoogleMeet: resultSync.wantGoogleMeet === false ? false : true,
-            wantTeams: resultSync.wantTeams === true ? true : false,
-            wantZoom: resultSync.wantZoom === true ? true : false
-        }, function () {
-            // Re-register content scripts whenever extension is installed or updated, provided permissions are available. Suppress notification for silent background operation.
-            reRegisterContentScripts()
-        })
-
-        checkAndCreateAlarm()
-    })
-})
-
-chrome.alarms.onAlarm.addListener((alarm) => {
-    if (alarm.name === ALARM_NAME) {
-        console.log("Alarm triggered: Running daily check...")
-        // checkPermissionsAndOpenMeetingsPage()
-    }
-})
-
-// Download transcripts, post webhook if URL is enabled and available
-// Fails if transcript is empty or webhook request fails or if no meetings in storage
-/** @throws error codes: 009, 010, 011, 012, 013, 014 */
-function processLastMeeting() {
-    return new Promise((resolve, reject) => {
-        pickupLastMeetingFromStorage()
-            .then(() => {
-                chrome.storage.local.get(["meetings"], function (resultLocalUntyped) {
-                    const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-                    chrome.storage.sync.get(["webhookUrl", "autoPostWebhookAfterMeeting", "autoDownloadFileAfterMeeting"], function (resultSyncUntyped) {
-                        const resultSync = /** @type {ResultSync} */ (resultSyncUntyped)
-
-                        // Create an array of promises to execute in parallel
-                        /** @type {Promise<any>[]} */
-                        const promises = []
-
-                        // Meeting index to download and post webhook
-                        // @ts-ignore - Because this line exists in the resolved promise from pickupLastMeetingFromStorage, which clearly means that at least one meeting exists and resultLocal.meetings cannot be undefined.
-                        const lastIndex = resultLocal.meetings.length - 1
-
-                        // Promise to download transcript
-                        if (resultSync.autoDownloadFileAfterMeeting) {
-                            promises.push(
-                                downloadTranscript(lastIndex)
-                            )
-                        }
-
-                        // Promise to post webhook if enabled
-                        if (resultSync.autoPostWebhookAfterMeeting && resultSync.webhookUrl) {
-                            promises.push(postTranscriptToWebhook(lastIndex))
-                        }
-
-                        // Execute all promises in parallel
-                        Promise.all(promises)
-                            .then(() => {
-                                resolve("Meeting processing and download/webhook posting complete")
-                                // Increment anonymous transcript generated count to a Google sheet
-                                // @ts-ignore - Because this line exists in the resolved promise from pickupLastMeetingFromStorage, which clearly means that at least one meeting exists and resultLocal.meetings cannot be undefined.
-                                const meetingSoftware = resultLocal.meetings[lastIndex].meetingSoftware
-                                const isWebhookEnabled = resultSync.webhookUrl && resultSync.autoPostWebhookAfterMeeting ? true : false
-                                fetch(`https://script.google.com/macros/s/AKfycbxK3Xd3u7ArtSEEUmu4jJSuJiyOHr5BtRqRGbcAy8Yc3zlVBTUoJ4wr-fJtQtlqRLGspg/exec?version=${chrome.runtime.getManifest().version}&isWebhookEnabled=${isWebhookEnabled}&meetingSoftware=${meetingSoftware}`, {
-                                    mode: "no-cors"
-                                })
-                            })
-                            .catch(error => {
-                                // Fails with error codes: 009, 010, 011, 012
-                                const parsedError = /** @type {ErrorObject} */ (error)
-                                console.error("Operation failed:", parsedError.errorMessage)
-                                reject({ errorCode: parsedError.errorCode, errorMessage: parsedError.errorMessage })
-                            })
-                    })
-                })
-            })
-            .catch((error) => {
-                // Fails with error codes: 013, 014
-                const parsedError = /** @type {ErrorObject} */ (error)
-                reject({ errorCode: parsedError.errorCode, errorMessage: parsedError.errorMessage })
-            })
-    })
-}
-
-/**
- * @throws error codes: 013, 014
- */
-// Process transcript and chat messages of the meeting that just ended from storage, format them into strings, and save as a new entry in meetings (keeping last 10)
-function pickupLastMeetingFromStorage() {
-    return new Promise((resolve, reject) => {
-        chrome.storage.local.get([
-            "meetingSoftware",
-            "meetingTitle",
-            "meetingStartTimestamp",
-            "transcript",
-            "chatMessages",
-        ], function (resultUntyped) {
-            const result = /** @type {ResultLocal} */ (resultUntyped)
-
-            if (result.meetingStartTimestamp) {
-                if ((result.transcript.length > 0) || (result.chatMessages.length > 0)) {
-                    // Create new transcript entry
-                    /** @type {Meeting} */
-                    const newMeetingEntry = {
-                        meetingSoftware: result.meetingSoftware ? result.meetingSoftware : "",
-                        meetingTitle: result.meetingTitle,
-                        meetingStartTimestamp: result.meetingStartTimestamp,
-                        meetingEndTimestamp: new Date().toISOString(),
-                        transcript: result.transcript,
-                        chatMessages: result.chatMessages,
-                        webhookPostStatus: "new"
-                    }
-
-                    // Get existing recent meetings and add the new meeting
-                    chrome.storage.local.get(["meetings"], function (resultLocalUntyped) {
-                        const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-                        let meetings = resultLocal.meetings || []
-                        meetings.push(newMeetingEntry)
-
-                        // Keep only last 10 transcripts
-                        if (meetings.length > 10) {
-                            meetings = meetings.slice(-10)
-                        }
-
-                        // Save updated recent transcripts
-                        chrome.storage.local.set({ meetings: meetings }, function () {
-                            console.log("Last meeting picked up")
-                            resolve("Last meeting picked up")
-                        })
-                    })
-                }
-                else {
-                    reject({ errorCode: "014", errorMessage: "Empty transcript and empty chatMessages" })
-                }
-            }
-            else {
-                reject({ errorCode: "013", errorMessage: "No meetings found. May be attend one?" })
-            }
-        })
-    })
-}
-
-
-
-/**
- * @param {number} index
- * @param {boolean} [isWebhookEnabled]
- * @throws error codes: 009, 010
- */
-function downloadTranscript(index, isWebhookEnabled) {
-    return new Promise((resolve, reject) => {
-        chrome.storage.local.get(["meetings"], function (resultLocalUntyped) {
-            const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-
-            if (resultLocal.meetings && resultLocal.meetings[index]) {
-                const meeting = resultLocal.meetings[index]
-
-                // Sanitise meeting title to prevent invalid file name errors
-                // https://stackoverflow.com/a/78675894
-                const invalidFilenameRegex = /[:?"*<>|~/\\\u{1}-\u{1f}\u{7f}\u{80}-\u{9f}\p{Cf}\p{Cn}]|^[.\u{0}\p{Zl}\p{Zp}\p{Zs}]|[.\u{0}\p{Zl}\p{Zp}\p{Zs}]$|^(CON|PRN|AUX|NUL|COM[1-9]|LPT[1-9])(?=\.|$)/gui
-                let sanitisedMeetingTitle = "Meeting"
-                if (meeting.meetingTitle) {
-                    sanitisedMeetingTitle = meeting.meetingTitle.replaceAll(invalidFilenameRegex, "_")
-                }
-                else if (meeting.title) {
-                    sanitisedMeetingTitle = meeting.title.replaceAll(invalidFilenameRegex, "_")
-                }
-
-                // Format timestamp for human-readable filename and sanitise to prevent invalid filenames
-                const timestamp = new Date(meeting.meetingStartTimestamp)
-                const formattedTimestamp = timestamp.toLocaleString("default", timeFormat).replace(/[\/:]/g, "-")
-
-                const prefix = meeting.meetingSoftware ? `${meeting.meetingSoftware} transcript` : "Transcript"
-
-                const fileName = `TranscripTonic/${prefix}-${sanitisedMeetingTitle} at ${formattedTimestamp} on.txt`
-
-
-                // Format transcript and chatMessages content
-                let content = getTranscriptString(meeting.transcript)
-                content += `\n\n---------------\nCHAT MESSAGES\n---------------\n\n`
-                content += getChatMessagesString(meeting.chatMessages)
-
-                // Add branding
-                content += "\n\n---------------\n"
-                content += "Transcript saved using TranscripTonic Chrome extension (https://chromewebstore.google.com/detail/ciepnfnceimjehngolkijpnbappkkiag)"
-                content += "\n---------------"
-
-                const blob = new Blob([content], { type: "text/plain" })
-
-                // Read the blob as a data URL
-                const reader = new FileReader()
-
-                // Read the blob
-                reader.readAsDataURL(blob)
-
-                // Download as text file, once blob is read
-                reader.onload = function (event) {
-                    if (event.target?.result) {
-                        const dataUrl = event.target.result
-
-                        // Create a download with Chrome Download API
-                        chrome.downloads.download({
-                            // @ts-ignore
-                            url: dataUrl,
-                            filename: fileName,
-                            conflictAction: "uniquify"
-                        }).then(() => {
-                            console.log("Transcript downloaded")
-                            resolve("Transcript downloaded successfully")
-                        }).catch((err) => {
-                            console.error(err)
-                            chrome.downloads.download({
-                                // @ts-ignore
-                                url: dataUrl,
-                                filename: "TranscripTonic/Transcript.txt",
-                                conflictAction: "uniquify"
-                            })
-                            console.log("Invalid file name. Transcript downloaded to TranscripTonic directory with simple file name.")
-                            resolve("Transcript downloaded successfully with default file name")
-
-                            // Logs anonymous errors to a Google sheet for swift debugging   
-                            fetch(`https://script.google.com/macros/s/AKfycbz0VFUYIke1WK12Q-8y-zQ91bOPRZ8dAL4cRpm309IYZO0k6uYDkTSlfbWFaGvUV_Z-JQ/exec?version=${chrome.runtime.getManifest().version}&code=009&error=${encodeURIComponent(err)}&meetingSoftware=${meeting.meetingSoftware}`, { mode: "no-cors" })
-                        })
-                    }
-                    else {
-                        reject({ errorCode: "009", errorMessage: "Failed to read blob" })
-                    }
-                }
-            }
-            else {
-                reject({ errorCode: "010", errorMessage: "Meeting at specified index not found" })
-            }
-        })
-    })
-}
-
-/**
- * @param {number} index
- * @throws error code: 010, 011, 012
- */
-function postTranscriptToWebhook(index) {
-    return new Promise((resolve, reject) => {
-        // Get webhook URL and meetings
-        chrome.storage.local.get(["meetings"], function (resultLocalUntyped) {
-            const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-            chrome.storage.sync.get(["webhookUrl", "webhookBodyType"], function (resultSyncUntyped) {
-                const resultSync = /** @type {ResultSync} */ (resultSyncUntyped)
-
-                if (resultSync.webhookUrl) {
-                    if (resultLocal.meetings && resultLocal.meetings[index]) {
-                        const meeting = resultLocal.meetings[index]
-
-                        /** @type {WebhookBody} */
-                        let webhookData
-                        if (resultSync.webhookBodyType === "advanced") {
-                            webhookData = {
-                                webhookBodyType: "advanced",
-                                meetingSoftware: meeting.meetingSoftware ? meeting.meetingSoftware : "",
-                                meetingTitle: meeting.meetingTitle || meeting.title || "",
-                                meetingStartTimestamp: new Date(meeting.meetingStartTimestamp).toISOString(),
-                                meetingEndTimestamp: new Date(meeting.meetingEndTimestamp).toISOString(),
-                                transcript: meeting.transcript,
-                                chatMessages: meeting.chatMessages
-                            }
-                        }
-                        else {
-                            webhookData = {
-                                webhookBodyType: "simple",
-                                meetingSoftware: meeting.meetingSoftware ? meeting.meetingSoftware : "",
-                                meetingTitle: meeting.meetingTitle || meeting.title || "",
-                                meetingStartTimestamp: new Date(meeting.meetingStartTimestamp).toLocaleString("default", timeFormat).toUpperCase(),
-                                meetingEndTimestamp: new Date(meeting.meetingEndTimestamp).toLocaleString("default", timeFormat).toUpperCase(),
-                                transcript: getTranscriptString(meeting.transcript),
-                                chatMessages: getChatMessagesString(meeting.chatMessages)
-                            }
-                        }
-
-                        // Post to webhook
-                        fetch(resultSync.webhookUrl, {
-                            method: "POST",
-                            headers: {
-                                "Content-Type": "application/json"
-                            },
-                            body: JSON.stringify(webhookData)
-                        }).then(response => {
-                            if (!response.ok) {
-                                throw new Error(`Webhook request failed with HTTP status code ${response.status} ${response.statusText}`)
-                            }
-                        }).then(() => {
-                            // Update success status.
-                            // @ts-ignore - Pointless type error about resultLocal.meetings being undefined, which is already checked above.
-                            resultLocal.meetings[index].webhookPostStatus = "successful"
-                            chrome.storage.local.set({ meetings: resultLocal.meetings }, function () {
-                                resolve("Webhook posted successfully")
-                            })
-                        }).catch(error => {
-                            console.error(error)
-                            // Update failure status.
-                            // @ts-ignore - Pointless type error about resultLocal.meetings being undefined, which is already checked above.
-                            resultLocal.meetings[index].webhookPostStatus = "failed"
-                            chrome.storage.local.set({ meetings: resultLocal.meetings }, function () {
-                                // Create notification and open webhooks page
-                                chrome.notifications.create({
-                                    type: "basic",
-                                    iconUrl: "icon.png",
-                                    title: "Could not post webhook!",
-                                    message: "Click to view status and retry. Check console for more details."
-                                }, function (notificationId) {
-                                    // Handle notification click
-                                    chrome.notifications.onClicked.addListener(function (clickedNotificationId) {
-                                        if (clickedNotificationId === notificationId) {
-                                            chrome.tabs.create({ url: "meetings.html" })
-                                        }
-                                    })
-                                })
-                                reject({ errorCode: "011", errorMessage: error })
-                            })
-                        })
-                    }
-                    else {
-                        reject({ errorCode: "010", errorMessage: "Meeting at specified index not found" })
-                    }
-                }
-                else {
-                    reject({ errorCode: "012", errorMessage: "No webhook URL configured" })
-                }
-            })
-        })
-    })
-}
-
-
-/**
- * Format transcript entries into string
- * @param {TranscriptBlock[] | []} transcript
- */
-function getTranscriptString(transcript) {
-    let transcriptString = ""
-    if (transcript.length > 0) {
-        transcript.forEach(transcriptBlock => {
-            transcriptString += `${transcriptBlock.personName} (${new Date(transcriptBlock.timestamp).toLocaleString("default", timeFormat).toUpperCase()})\n`
-            transcriptString += transcriptBlock.transcriptText
-            transcriptString += "\n\n"
-        })
-        return transcriptString
-    }
-    return transcriptString
-}
-
-
-/**
- * Format chat messages into string
- * @param {ChatMessage[] | []} chatMessages
- */
-function getChatMessagesString(chatMessages) {
-    let chatMessagesString = ""
-    if (chatMessages.length > 0) {
-        chatMessages.forEach(chatMessage => {
-            chatMessagesString += `${chatMessage.personName} (${new Date(chatMessage.timestamp).toLocaleString("default", timeFormat).toUpperCase()})\n`
-            chatMessagesString += chatMessage.chatMessageText
-            chatMessagesString += "\n\n"
-        })
-    }
-    return chatMessagesString
-}
-
-function clearTabIdAndApplyUpdate() {
-    // Nullify to indicate end of meeting processing
-    chrome.storage.local.set({ meetingTabId: null }, function () {
-        console.log("Meeting tab id cleared for next meeting")
-
-        // Check if there's a deferred update
-        chrome.storage.local.get(["isDeferredUpdatedAvailable"], function (resultLocalUntyped) {
-            const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-
-            if (resultLocal.isDeferredUpdatedAvailable) {
-                console.log("Applying deferred update")
-                chrome.storage.local.set({ isDeferredUpdatedAvailable: false }, function () {
-                    chrome.runtime.reload()
-                })
-            }
-        })
-    })
-}
-
-/** @throws error codes: 009, 010, 011, 012, 013, 014 */
-function recoverLastMeeting() {
-    return new Promise((resolve, reject) => {
-        chrome.storage.local.get(["meetings", "meetingStartTimestamp"], function (resultLocalUntyped) {
-            const resultLocal = /** @type {ResultLocal} */ (resultLocalUntyped)
-            // Check if user ever attended a meeting
-            if (resultLocal.meetingStartTimestamp) {
-                /** @type {Meeting | undefined} */
-                let lastSavedMeeting
-                if ((resultLocal.meetings) && (resultLocal.meetings.length > 0)) {
-                    lastSavedMeeting = resultLocal.meetings[resultLocal.meetings.length - 1]
-                }
-
-                // Last meeting was not processed for some reason. Need to recover that data, process and download it.
-                if ((!lastSavedMeeting) || (resultLocal.meetingStartTimestamp !== lastSavedMeeting.meetingStartTimestamp)) {
-                    processLastMeeting().then(() => {
-                        resolve("Recovered last meeting to the best possible extent")
-                    }).catch((error) => {
-                        // Fails with error codes: 009, 010, 011, 013, 014
-                        const parsedError = /** @type {ErrorObject} */ (error)
-                        reject({ errorCode: parsedError.errorCode, errorMessage: parsedError.errorMessage })
-                    })
-                }
-                else {
-                    resolve("No recovery needed")
-                }
-            }
-            else {
-                reject({ errorCode: "013", errorMessage: "No meetings found. May be attend one?" })
-            }
-        })
-    })
-}
-
-/**
- * @param {string|string[]} platform
- */
-function getPermissionStatus(platform) {
-    /**
-     * @param {string} p
-     * @returns {Promise<string>}
-     */
-    function fetchSinglePlatformStatus(p) {
-        return new Promise((resolve, reject) => {
-            const config = PLATFORM_CONFIGS[p]
-
-            if (!config) {
-                reject(`Invalid platform: ${p}`)
-                return
-            }
-
-            chrome.permissions.contains({
-                origins: config.matches,
-                permissions: config.permissions
-            }).then(function (hasPermission) {
-                if (hasPermission) {
-                    resolve("Enabled")
-                } else {
-                    resolve("Disabled")
-                }
-            })
-        })
-    }
-
-    if (Array.isArray(platform)) {
-        // Multi-platform path
-        return Promise.all(platform.map(function (p) {
-            return fetchSinglePlatformStatus(p)
-        }))
-    } else {
-        // Single platform path
-        return fetchSinglePlatformStatus(platform)
-    }
-}
-
-
-/**
- * @param {Platform|Platform[]} platform
- */
-function requestPlatformPermission(platform) {
-    return new Promise((resolve, reject) => {
-        const platforms = Array.isArray(platform) ? platform : [platform]
-
-        // Map each platform to its config, then flatten the resulting nested arrays
-        const allOrigins = platforms.flatMap(p => PLATFORM_CONFIGS[p]?.matches || [])
-        const allPermissions = platforms.flatMap(p => PLATFORM_CONFIGS[p]?.permissions || [])
-
-        // Use Set to handle duplicates
-        const request = {
-            origins: [...new Set(allOrigins)],
-            permissions: [...new Set(allPermissions)]
-        }
-
-        chrome.permissions.request(request)
-            .then((granted) => {
-                granted ? resolve("Permissions granted") : reject("Permissions denied")
-            })
-            .catch((error) => {
-                console.error(error)
-                reject(`Could not get permissions for ${platforms.join(", ")}`)
-            })
-    })
-}
-
-/**
- * @param {Platform|Platform[]} platform
- */
-function getContentScriptStatus(platform) {
-    /**
-    * @param {string} p
-    * @returns {Promise<string>}
-    */
-    function getSinglePlatformContentScriptStatus(p) {
-        return new Promise((resolve, reject) => {
-            const config = PLATFORM_CONFIGS[p]
-
-            if (!config) {
-                reject(`Invalid platform: ${p}`)
-                return
-            }
-
-            chrome.scripting
-                .getRegisteredContentScripts()
-                .then((scripts) => {
-                    const isRegistered = scripts.some(s => s.id === config.id)
-
-                    if (isRegistered) {
-                        resolve("Enabled")
-                    } else {
-                        resolve("Disabled")
-                    }
-                })
-                .catch((error) => {
-                    console.error(`Error fetching status for ${p}:`, error)
-                    reject(`Could not retrieve status for ${p}`)
-                })
-        })
-    }
-
-    if (Array.isArray(platform)) {
-        // Multi-platform path
-        return Promise.all(platform.map(function (p) {
-            return getSinglePlatformContentScriptStatus(p)
-        }))
-    } else {
-        // Single platform path
-        return getSinglePlatformContentScriptStatus(platform)
-    }
-}
-
-/**
- * @param {Platform|Platform[]} platform
- */
-function registerContentScript(platform, showNotification = true) {
-    /**
-     * @param {string} p
-     * @returns {Promise<string>}
-     */
-    function registerSinglePlatformContentScript(p) {
-        return new Promise((resolve, reject) => {
-            const config = PLATFORM_CONFIGS[p]
-
-            // Map each match to a permission check
-            const permissionChecks = config.matches.map(pattern =>
-                chrome.permissions.contains({ origins: [pattern] }).then(result => result ? pattern : null)
-            )
-
-            Promise.all(permissionChecks).then((results) => {
-                // Filter to get only patterns with existing permissions
-                const allowedMatches = results.filter(pattern => pattern !== null)
-
-                if (allowedMatches.length > 0) {
-                    chrome.scripting
-                        .getRegisteredContentScripts()
-                        .then((scripts) => {
-                            let isRegistered = scripts.some(s => s.id === config.id)
-
-                            if (isRegistered) {
-                                console.log(`${p} content script already registered`)
-                                resolve(`Content script already registered`)
-                            } else {
-                                chrome.scripting.registerContentScripts([{
-                                    id: config.id,
-                                    js: config.js,
-                                    matches: allowedMatches,
-                                    excludeMatches: config.excludeMatches,
-                                    runAt: "document_end",
-                                }])
-                                    .then(() => {
-                                        console.log(`${p} content script registered successfully.`)
-
-                                        if (showNotification) {
-                                            chrome.permissions.contains({
-                                                permissions: ["notifications"]
-                                            }).then((hasNotifyPermission) => {
-                                                if (hasNotifyPermission) {
-                                                    chrome.notifications.create({
-                                                        type: "basic",
-                                                        iconUrl: "icon.png",
-                                                        title: "Enabled!",
-                                                        message: p === "google_meet" ? `Refresh any existing meeting pages` : ` ${p === "teams" ? `Join Teams meetings on the browser` : `Zoom meetings will automatically open in the browser`}. Refresh any existing pages.`
-                                                    })
-                                                }
-                                            })
-                                        }
-                                        resolve(`Content script registered`)
-                                    })
-                                    .catch((error) => {
-                                        console.error(`${p} registration failed.`, error)
-                                        reject(`Failed to register content script`)
-                                    })
-                            }
-                        })
-                } else {
-                    reject(`Insufficient permissions`)
-                }
-            })
-        })
-    }
-
-    if (Array.isArray(platform)) {
-        // Multi-platform path
-        return Promise.all(platform.map(function (p) {
-            return registerSinglePlatformContentScript(p)
-        }))
-    } else {
-        // Single platform path
-        return registerSinglePlatformContentScript(platform)
-    }
-}
-
-/**
- * @param {Platform|Platform[]} platform
- */
-function deregisterContentScript(platform) {
-    /**
-    * @param {string} p
-    * @returns {Promise<string>}
-    */
-    function deregisterSinglePlatformContentScript(p) {
-        return new Promise((resolve, reject) => {
-            const config = PLATFORM_CONFIGS[p]
-
-            chrome.scripting
-                .getRegisteredContentScripts()
-                .then((scripts) => {
-                    let isRegistered = scripts.some(s => s.id === config.id)
-
-                    if (!isRegistered) {
-                        console.log(`${p} content script not registered`)
-                        resolve(`Content script not registered`)
-                    } else {
-                        chrome.scripting.unregisterContentScripts({
-                            ids: [config.id]
-                        })
-                            .then(() => {
-                                console.log(`${p} content script deregistered successfully.`)
-                                resolve(`Content script deregistered`)
-                            })
-                            .catch((error) => {
-                                console.error(`${p} deregistration failed.`, error)
-                                reject(`Failed to deregister content script`)
-                            })
-                    }
-                })
-        })
-    }
-
-    if (Array.isArray(platform)) {
-        // Multi-platform path
-        return Promise.all(platform.map(function (p) {
-            return deregisterSinglePlatformContentScript(p)
-        }))
-    } else {
-        // Single platform path
-        return deregisterSinglePlatformContentScript(platform)
-    }
-}
-
-
-function reRegisterContentScripts() {
-    chrome.storage.sync.get(["wantGoogleMeet", "wantTeams", "wantZoom"], function (resultSyncUntyped) {
-        const resultSync = /** @type {ResultSync} */ (resultSyncUntyped)
-
-        /** @type {Promise<any>[]} */
-        const reRegisterPromises = []
-
-        /** @type {Platform[]} */
-        const wantedPlatforms = []
-
-        // Register content scripts if permissions are available and if user has not explicitly opted out
-        if (resultSync.wantGoogleMeet) {
-            wantedPlatforms.push("google_meet")
-        }
-        if (resultSync.wantTeams) {
-            wantedPlatforms.push("teams")
-        }
-        if (resultSync.wantZoom) {
-            wantedPlatforms.push("zoom")
-            reRegisterPromises.push(registerZoomRedirect())
-        }
-
-        if (wantedPlatforms.length > 0) {
-            reRegisterPromises.push(registerContentScript(wantedPlatforms, false))
-        }
-
-        Promise.all(reRegisterPromises)
-            .then(() => {
-                console.log("Permitted scripts re-registered successfully.")
-            })
-            .catch((error) => {
-                console.log(error)
-            })
-    })
-}
-
-
-function registerZoomRedirect() {
-    return new Promise((resolve, reject) => {
-        // Check if we have the host permission required for the redirect
-        chrome.permissions.contains({
-            origins: ["https://*.zoom.us/*"]
-        }).then((hasHostPermission) => {
-            if (hasHostPermission) {
-                // Check if the ruleset is already enabled to avoid redundant updates
-                chrome.declarativeNetRequest.getEnabledRulesets().then((enabledRulesets) => {
-                    const rulesetId = "ruleset_1"
-
-                    if (enabledRulesets.includes(rulesetId)) {
-                        console.log("Zoom redirect ruleset already active")
-                        resolve("Zoom redirect already active")
-                    }
-                    else {
-                        // Enable the ruleset
-                        chrome.declarativeNetRequest.updateEnabledRulesets({
-                            enableRulesetIds: [rulesetId]
-                        }).then(() => {
-                            console.log("Zoom redirect ruleset enabled successfully")
-                            resolve("Zoom redirect registered")
-                        }).catch((error) => {
-                            console.error("Failed to enable DNR ruleset:", error)
-                            reject("Failed to enable redirect ruleset")
-                        })
-                    }
-                })
-            }
-            else {
-                reject("Insufficient permissions")
-                return
-            }
-        })
-    })
-}
-
-function deregisterZoomRedirect() {
-    return new Promise((resolve, reject) => {
-        const rulesetId = "ruleset_1"
-
-        chrome.declarativeNetRequest.getEnabledRulesets().then((enabledRulesets) => {
-            if (!enabledRulesets.includes(rulesetId)) {
-                console.log("Zoom redirect ruleset already disabled")
-                resolve("Zoom redirect already disabled")
-            }
-            else {
-                // Disable the ruleset
-                chrome.declarativeNetRequest.updateEnabledRulesets({
-                    disableRulesetIds: [rulesetId]
-                }).then(() => {
-                    console.log("Zoom redirect ruleset disabled successfully")
-                    resolve("Zoom redirect deregistered")
-                }).catch((error) => {
-                    console.error("Failed to disable DNR ruleset:", error)
-                    reject("Failed to disable redirect ruleset")
-                })
-            }
-        })
-    })
-}
-
-/**
- * Opens the extension popup programmatically.
- */
-function openExtensionPopup() {
-    return new Promise((resolve, reject) => {
-        chrome.action.openPopup()
-            .then(() => {
-                console.log("Popup opened successfully")
-                resolve("Popup opened")
-            })
-            .catch((error) => {
-                console.error("Failed to open popup:", error)
-                reject("Failed to open popup")
-            })
-    })
-}
-
-function checkPermissionsAndOpenMeetingsPage() {
-    console.log("Check permissions")
-    chrome.storage.sync.get(["wantGoogleMeet", "wantTeams", "wantZoom"], function (resultSyncUntyped) {
-        const resultSync = /** @type {ResultSync} */ (resultSyncUntyped)
-
-        /** @type {Platform[]} */
-        const wantedPlatforms = []
-        //  Consider enabled if user has not explicitly opted out
-        if (resultSync.wantGoogleMeet) {
-            wantedPlatforms.push("google_meet")
-        }
-        if (resultSync.wantTeams) {
-            wantedPlatforms.push("teams")
-        }
-        if (resultSync.wantZoom) {
-            wantedPlatforms.push("zoom")
-        }
-
-        /** @type {ExtensionMessage} */
-        const message = {
-            type: "get_platform_permission_status",
-            platform: wantedPlatforms
-        }
-        getPermissionStatus(wantedPlatforms).then((result) => {
-            console.log(result)
-
-            /** @type {Platform[]} */
-            const permissionMissingPlatforms = []
-
-            for (let i = 0; i < wantedPlatforms.length; i++) {
-                if (Array.isArray(result) && result[i] === "Disabled") {
-                    permissionMissingPlatforms.push(wantedPlatforms[i])
-                }
-            }
-
-            if (permissionMissingPlatforms.length > 0) {
-                console.log(permissionMissingPlatforms)
-                chrome.tabs.create({
-                    url: chrome.runtime.getURL("meetings.html")
-                })
-            }
-        })
-    })
-}
-
-function checkAndCreateAlarm() {
-    chrome.alarms.get(ALARM_NAME, (alarm) => {
-        if (!alarm) {
-            chrome.alarms.create(ALARM_NAME, {
-                periodInMinutes: INTERVAL_IN_MINUTES
-            })
-            console.log(`Alarm "${ALARM_NAME}" created successfully.`)
-        } else {
-            console.log(`Alarm "${ALARM_NAME}" already exists. Next scheduled run:`, new Date(alarm.scheduledTime))
-        }
-    })
-}
